feat(diff-panel): make preview tabs draggable into Claude PTY (#12)
File and editor tabs in the diff panel's TabStrip now drag the same way
file-tree rows already do — pointer-based, ghost element follows the
cursor, drop on a Claude or shell PTY publishes the file path through
INTERNAL_DROP_EVENT, App.tsx receives it and writes the @rel-formatted
payload into the target terminal.

Closes the workflow loop the Cmd+K palette opened: ⌘K → file lands in
preview → drag tab into Claude → continue typing. Before this, the
user had to bounce back to the sidebar tree to re-find the file.

Refactor: the ~110-line drag block previously lived inline in
tree-node.tsx. Extracted to `src/lib/use-internal-drag.ts` as
`createInternalDrag(source)` — caller-supplied accessor returns
`{ path, label } | null` (null means "this row/tab isn't draggable",
e.g. the Git changes tab). The hook owns threshold detection, ghost
rendering, hover-target tracking, drop dispatch, and the
`consumedClick()` accessor that lets the click handler bail when a
gesture turned into a drag. tree-node.tsx is back to ~140 lines and
the new TabItem usage is ~15 lines.

No new dependencies. No PTY, Rust, or context changes — purely UI
plus one new util file.
